Prueba 2 Ejercicio 2

+++ src/exercises.cpp
@@ -20,7 +20,7 @@ void exercise_2(double A, double B, double C) {
   
   }
 
-void exercise_3(int a, int b) {
+void exercise_3(int a, float b) {
   int result = a / b;
   if (b == 0) {
       cout << "Imposible";
